requirements.txt
#hermes bindings
hermes-python>=0.1

#python-scriptures
python-scriptures
